# specs/openapi.yaml
openapi: 3.1.0
info:
  title: NarrativeOS API
  version: 0.3.0
paths:
...
  /v1/sessions:
    get:
      summary: List sessions
      responses:
        "200":
          description: Session list
    post:
      summary: Create a new narrative session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [world_id, initial_state]
              properties:
                world_id:
                  type: string
                initial_state:
                  $ref: "./narrative_state.schema.json"
                player_profile:
                  type: object
                  additionalProperties: true
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Session created
...
  /v1/sessions/{session_id}/replay:
    get:
      summary: Replay a session with event, critic, and state traces
      parameters:
        - in: path
          name: session_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Replay payload
...
  /v1/library/worlds/{world_id}:
    get:
      summary: Reader world detail with published version and manifest
      parameters:
        - in: path
          name: world_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: World detail payload
  /v1/reader/sessions:
    post:
      summary: Create a reader session against the latest published world version
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [world_id]
              properties:
                world_id:
                  type: string
                reader_id:
                  type: string
      responses:
        "200":
          description: Reader session created
  /v1/reader/continue:
    post:
      summary: Continue a story with entitlement and metering checks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id]
              properties:
                session_id:
                  type: string
                account_id:
                  type: string
                choice_id:
                  type: string
                freeform_intent:
                  type: string
                reader_id:
                  type: string
      responses:
        "200":
          description: Reader chapter view or paywall response
  /v1/reader/sessions/{session_id}/quote:
    get:
      summary: Quote continue cost / access tier for the next chapter
      parameters:
        - in: path
          name: session_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Entitlement quote
  /v1/reader/sessions/{session_id}/prefill:
    get:
      summary: Build next-intent prefill for the sticky composer
      parameters:
        - in: path
          name: session_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Intent prefill payload
...
  /v1/reader/checkout/start:
    post:
      summary: Start a web-first checkout stub session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tier_id]
              properties:
                account_id:
                  type: string
                reader_id:
                  type: string
                tier_id:
                  type: string
                provider:
                  type: string
      responses:
        "200":
          description: Checkout session stub
...